% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_read.R
\name{mt_example}
\alias{mt_example}
\title{Get path to \code{move2} example data}
\usage{
mt_example()
}
\value{
The path to the example file of the \code{move2} package
}
\description{
The \code{move2} package comes with an example data file that is directly downloaded from
\href{https://www.movebank.org}{movebank} and gz compressed for reduction in package size.
}
\details{
The example dataset is the study "Martes pennanti LaPoint New York" (study id: \code{69258089}), shared under
the CC-BY-NC license.
For more information on the data see LaPoint et al. (2013) Landscape Ecology. doi: 10.1007/s10980-013-9910-0
}
\examples{

# Get path to one example
mt_example()
mt_read(mt_example())
}
\seealso{
\itemize{
\item \code{\link[=mt_read]{mt_read()}} to read in the file
}
}
